Estilos

+++ src/app/carrito/carrito/carrito.jsx
@@ -2,7 +2,7 @@
 import { use, useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-import { getAllCarts, removeCartBack } from "@/redux/actions";
+import { AddCartBack, getAllCarts, removeCartBack } from "@/redux/actions";
 
 import NavBar from "@/components/navbar/Navbar";
 import Link from "next/link";
@@ -11,10 +11,12 @@ import styles from "../carrito/carrito.module.css";
 import { typographyClasses } from "@mui/material";
 
 import toast, { Toaster } from "react-hot-toast";
+import Swal from "sweetalert2";
 
 const Carrito = () => {
   const dispatch = useDispatch();
   const carts = useSelector((state) => state.carrito);
+  const [restored, setRestored] = useState([]);
   const [refresh, setRefresh] = useState(false);
   const [totalPrice, setTotalPrice] = useState(0);
   const [client, setClient] = useState(false);
@@ -37,11 +39,44 @@ const Carrito = () => {
   useEffect(() => {
     loadCarts();
     calcularTotal();
-  }, [carts, refresh]);
+  }, [carts, refresh, restored]);
+
+  const restoreCart = async () => {
+    if (restored.length > 0) {
+      console.log(restored);
+      await Promise.all(
+        restored.map(async (cart) => {
+          console.log("soy producto", cart);
+          await dispatch(AddCartBack({ UserId: user.id, ProductId: cart }));
+        })
+      );
+      toast.success("Restored");
+      setRestored([]);
+      setRefresh(true);
+    } else
+      toast("Nothing to restore", {
+        duration: 750,
+      });
+  };
 
   const handleCarrito = async (id) => {
-    await dispatch(removeCartBack({ UserId: user.id, ProductId: id }));
-    setRefresh(true);
+    const deleted = carts.filter((fav) => fav.id === id);
+    Swal.fire({
+      title: "Do you want to delete this product from favorites?",
+      showDenyButton: true,
+      showCancelButton: true,
+      confirmButtonText: "Yes",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        setRestored([...restored, deleted]);
+        console.log("soy restore", restored);
+        await dispatch(removeCartBack({ UserId: user.id, ProductId: id }));
+        setRefresh(true);
+        Swal.fire("Deleted!", "", "success");
+      } else if (result.isDenied) {
+        Swal.fire("Ok", "", "info");
+      }
+    });
   };
 
   const calcularTotal = () => {
@@ -88,6 +123,7 @@ const Carrito = () => {
       {carts?.length > 0 ? (
         <>
           <h1 className={styles.title}>Shopping Cart</h1>
+
           <div className={styles.bigCont}>
             <div className={styles.otherCont}>
               <div>
@@ -124,18 +160,22 @@ const Carrito = () => {
                 ))}
               </div>
             </div>
-
-            <div className={styles.pay}>
-              <h3 className={styles.price}>{`TOTAL: ${totalPrice}`}</h3>
-              <br />
-              <button onClick={handleClick} className={styles.Btn}>
-                Pay
-                <svg viewBox="0 0 576 512" className={styles.svgIcon2}>
-                  <path
-                    className={styles.path}
-                    d="M512 80c8.8 0 16 7.2 16 16v32H48V96c0-8.8 7.2-16 16-16H512zm16 144V416c0 8.8-7.2 16-16 16H64c-8.8 0-16-7.2-16-16V224H528zM64 32C28.7 32 0 60.7 0 96V416c0 35.3 28.7 64 64 64H512c35.3 0 64-28.7 64-64V96c0-35.3-28.7-64-64-64H64zm56 304c-13.3 0-24 10.7-24 24s10.7 24 24 24h48c13.3 0 24-10.7 24-24s-10.7-24-24-24H120zm128 0c-13.3 0-24 10.7-24 24s10.7 24 24 24H360c13.3 0 24-10.7 24-24s-10.7-24-24-24H248z"
-                  ></path>
-                </svg>
+            <div className={styles.scndCont}>
+              <div className={styles.pay}>
+                <h3 className={styles.price}>{`TOTAL: ${totalPrice}`}</h3>
+                <br />
+                <button onClick={handleClick} className={styles.Btn}>
+                  Pay
+                  <svg viewBox="0 0 576 512" className={styles.svgIcon2}>
+                    <path
+                      className={styles.path}
+                      d="M512 80c8.8 0 16 7.2 16 16v32H48V96c0-8.8 7.2-16 16-16H512zm16 144V416c0 8.8-7.2 16-16 16H64c-8.8 0-16-7.2-16-16V224H528zM64 32C28.7 32 0 60.7 0 96V416c0 35.3 28.7 64 64 64H512c35.3 0 64-28.7 64-64V96c0-35.3-28.7-64-64-64H64zm56 304c-13.3 0-24 10.7-24 24s10.7 24 24 24h48c13.3 0 24-10.7 24-24s-10.7-24-24-24H120zm128 0c-13.3 0-24 10.7-24 24s10.7 24 24 24H360c13.3 0 24-10.7 24-24s-10.7-24-24-24H248z"
+                    ></path>
+                  </svg>
+                </button>
+              </div>
+              <button onClick={restoreCart} className={styles.rstrBtn}>
+                Restore
               </button>
             </div>
           </div>
